Add ability to save .mp4 files from URLs into the filesystem and notes about usage/next steps
If the file is saved locally, the seeking-not-working issue from the UrlResource proxy method is resolved since the system handles 'Ranges' header appropriately and loads super fast. However, as it is written currently, the file has to be downloaded before being served to the user, which is a bad experience. Thus, while having the .mp4 files locally is the best option, it doesn't result in the quickest experience. Next step: either use a streaming response or toggle between UrlResource proxy and using downloaded files