Add very base (and partially non-functional) Kissanime content for bypassing Cloudflare
The PageConfiguration takes forever to run, so it needs to be put on a different thread. As it stands, theoretically, the request to /searchKissanime should return the correct links; however, the PageConfiguration>Interceptor is needed to get the secure cookies. Once the async config is done, this should function as desired - to get the title/URL suggestions from the Kissanime search box