Improve kissanime cookie handling if auth cookie isn't found
Previously, the getAuthCookie() would simply return a blank cookie if the auth-cookie key wasn't found. However, if the cookie wasn't found, then waitForCloudflareToAllowAccessToKissanime() wouldn't have known because getAuthCookie().hasExpired() would be false. Thus, add a max-age of 0 to the placeholder cookie so that getAuthCookie().hasExpired() will be true if the cookie isn't found