Enhance KissanimeVideoHostRequest with captchaAnswer field and add respective response object
As explained in a previous commit, Kissanime no longer allows multiple requests for one single AreYouHuman question. As such, the user will have to fill out the answer themselves. In preparation for such logic, add the ability to pass the captcha answer in the same request body and add a response that has both the videoHostUrl field (for successful captcha bypasses) and captchaContent field (for when they need to fill out the captcha)