Ensure MP4 file length is larger than the 'Range' start-byte before serving file
MP4 file downloads are initiated upon NovelPlanet source resolution, but they might not be finished downloading by the time the user requests the video content. If this is the case, the video will be found but unable to be served. Thus, check the video file length before attempting to read it from disk; if it hasn't downloaded enough, then return a UrlResource with the backup URL instead