Replace Spring WebFlux with Spring Web
Not only is WebFlux unnecessary in this app, it ends up overwriting manual setting of headers which prevents simple proxying of data from third-party back to the user. It does this by overwriting any header set in the code with its own header based on the response body, e.g. if returning a part of a Resource and I set Content-Length: start-end/total, it will replace that Content-Length header with start-end/end. While this could be solved using interceptors, don't use those and instead simply replace WebFlux with Web; this is okay because WebFlux wasn't really used anywhere else in the app. Luckily, this simplifies everything because it gives the dev full control over all response content, including headers that don't necessarily match exactly what's being returned as well as not needing to manually route '/' to index.html (Web automatically serves any index.html file it finds in a 'public', 'static', or 'webapp' folder in resources/