Copy logic for getting Kissanime cookie to separate application
Heroku has a limit on the amount of RAM an application can use: 512 kb for the free, hobby, and standard-1x tiers; RAM isn't increased until you pay for the standard-2x or higher tier, which is an unreasonable 0/month. Thus, if the JavaFX portion of the app (which is used to get the cookie from Kissanime) is split out to a separate application (cookie-getter), it might be possible to lower the RAM usage of this app from 1-1.5 GB to < 512 kb. Further, after running cookie-getter separately, it appears to use about 500-700 kb of RAM, but Heroku hasn't killed the app, so hopefully this will be a sustainable way to run the app for free on Heroku