Move last-watched-episodes text from Episodes column to top of page
The Episodes column was slightly cluttered when it held both last-watched-episode text and the text pushed the scroll content down in an unpleasant way such that the Shows/Episodes scroll content weren't aligned nicely. Fix that by moving the last-watched-episodes text to the top of the page (as a sub-header). Also, make it show last-watched-episodes for each of the possible show options and only when 'Watch' tab is selected so it doesn't look out of place being at the top of the page