Make Shows/Episodes sections flow on top of each other on sm screens
On mobile/sm screen sizes, it was really difficult to read the show/episode titles when they were side-by-side. Fix this by making the Show/Episode sections flow one on top of the other on sm screen sizes. Similarly, update the episode-count-badge to display on all screen sizes because now that the Shows section is wider, it has enough space to hold the badge