Add top/bottom borders to Shows/Episodes scroll divs
When not scrolled all the way to the top/bottom of one of these lists, the top/bottom of Shows/Episodes divs disappears which blurs the line of where one begins and the next ends. This is especially the case in sm screens where the two are stacked on top of each other. Fix that by adding borders. Only add borders to the top/bottom since the individual .list-group-item elements already have left/right borders