Convert Dockerfile CMD from 'exec' form to 'shell' form
Heroku passes env vars to `docker run` through -e flag. For example, Heroku has a different value of PORT for every run and will pass it to containers through `docker run -e PORT=$PORT app`.

As explained in the [Dockerfile CMD docs](https://devcenter.heroku.com/articles/container-registry-and-runtime#get-the-port-from-the-environment-variable), CMD can be either run as a plain executable (array of strings, each being the equivalent to space-separated text in the terminal) or nested inside a shell (`sh`) run command (equivalent of `/bin/sh -c <CMD_line>`).

While Docker docs specify that 'exec' form is preferred, it causes issues when reading env vars in `docker run -e ENV_KEY=env_value container_name` in that if an env var is not defined *and* not passed as an arg to one of the options in the CMD line, it fails to run.

For example, given the line below in 'exec' form:
`[ "java", "${JAVA_OPTS}", "-jar", "spring-app.war", "--server.port=${PORT:-8080}" ]`
If PORT isn't defined or passed in `docker run -e ENV_KEY=env_value`, it will still run since it is passed as an arg to the CMD option `--server.port`.
However, if JAVA_OPTS isn't defined, `docker run` fails because JAVA_OPTS isn't passed as an arg to an option, but is instead intended to be spread as options to `java` (e.g. JAVA_OPTS=-Xmx640m -> `java -Xmx640m ...`). As such, if the variable is not defined *and* not passed as an arg to an option in CMD, the container cannot run.

This issue does not occur in 'shell' form because it is run inside an `sh` command like below, so if the variable isn't defined, it doesn't break the run. This is because shell will automatically just make JAVA_OPTS null (not empty string, which would make java try to run `''` giving file-doesn't-exist error), with the net effect of ignoring it.
`/bin/sh -c java ${JAVA_OPTS} -jar spring-app.war --server.port=${PORT:-8080}`

Finally, note that `${PORT:8080}` was changed to `${PORT:-8080}` since that defaults the value in shell lines (vs Docker CMD lines) but also doesn't set PORT after the command is run.