Fix double padding addition from App's .container-fluid .col
In Bootstrap, both .container(-fluid) and .col have a left/right padding of 15px, and .row counteracts that with a left/right margin of -15px. As such, it is expected that the first thing rendered in the .container(-fluid) element is a .row so that the extra padding of .container(-fluid) is erased by .row. However, the App rendered a .col-12 as the first thing in the .container-fluid element. Thus, replace the first-rendered div CSS class from .col-12 to .row; likewise, update the other components to have an updated .col as the first thing rendered instead of .row. Finally, preserve some of the extra/accidental padding in Show's .card content to give some extra space for mobile users to be able to scroll the whole Show page (not the list content) easier