Add Header in About page with 'Back' button for ease of navigation
It is possible some users wouldn't want to press the 'back' button in the browser and would prefer an explicit 'Back' button to click, so add it to the About page with the styling applied from Header.

Also, add PropTypes checking for the `history` object passed down from react-router-dom. Attempt to use that for the 'Back' button's `onClick` handler first, with the native `window.history.back()` function as a fallback.