Add try/catch to remaining Kissanime requests in case cookie is invalidated before expiration
Like what was done in commit effb74a6, occasionally, Kissanime will invalidate a cookie before it expires and the Cloudflare-bypass will have to be redone. Previously, only `/searchKissanime`/`searchKissanimeTitles()` was checked for the invalidated cookie but, since the response is cached, the other Kissanime endpoints need to have the same type of check in place.

TODO: Abstract this try/catch to a separate function that accepts a Consumer instead of duplicating all this code in multiple places.