Overload bypassCloudflareDdosScreen() to toggle throwing the RuntimeException
All Kissanime-scraping endpoints require the Cloudflare cookie to function, so they should still throw errors if cookie-extraction fails; that way, the user knows to refresh the page if they fail.

However, if the cookie-extraction fails on app boot, then the app crashes and won't recover. To avoid this, don't throw the error only on app boot so that it can at least start running and then subsequent endpoint requests can serve to try to extract the cookie again.