Add new .properties config for toggling Kissanime requests
Since Kissanime has been shut down, it is no longer able to be scraped. Rather than throw all that code away, use a config variable to toggle if it is available to be parsed. In the process, also convert the endpoints to return ResponseEntity<MyResponseObj> so that empty responses can be returned