Sort episodes by number in title rather than blind list reversal
Kissanime returned the episode list in descending order based on how the website was scraped. Since 4anime scraping gets a list in ascending order (which is what we want), the list reversal shows the last episode first. Fix that by not assuming the sort-order of the response and manually sorting by episode number