Add extra settings to the proxy so that the hosted website can be used
In the event that the local back-end server isn't running, the 'target' field can be replaced with the actual hosted website. This way, the front-end can be tested without having to run a local back-end server. To do so, the origin needs to be changed to be that of the website as well as disabling the HTTPS certificate checks since the local front-end server is clearly not certified