Move hard-coded string for `tools.jar` from comment to a fallback in the Gradle build logic
This way, even if our original dynamic code fails, the file will still be defined. Note: `/path/to/java/lib/tools.jar` is needed only for dev commands like `javac` and not for simply executing `.jar` files (which is why it's only just recently that it was discovered the file was needed in the first place).