Update GitHub workflows/actions to correctly cache files
Artifacts are for files you want to be relatively permanent, like deployment .tar files or Docker images. Caches are for temporary files you want to share between workflows/jobs/actions but not shared with the world. Env vars don't persist between jobs or workflows, so add a separate, shared action to set the env vars needed.