Bug fix: Preserve array indexing when forwarding args to underlying Bash functions
`$@` is a special variable that can be used as both a String and an Array, but if setting it to another variable directly (e.g. `myVar="$@"`), it loses the ability to behave as an Array. As such, to preserve indices when forwarding args to other functions from `main()`, we must set the args into an Array variable and forward the array contents manually.