Bug fix: Replace `"$@"` in `docker run` calls since it prevents using the default `CMD` from Dockerfile
Since our Dockerfile uses `CMD` to set the default command to run, passing anything (even an empty string caused by `"$@"` when the arg array is empty) to the `docker run` command will override `CMD`. Fix this by (1) splitting out the args via manual processing, and (2) unquotting the command so empty strings are ignored.

In the process of (1), enhance how `dockerRun()` is written so that it automatically adds `-e` to the generated Docker container's env vars, ports are defaulted (or removed if `dockerRun -p=` is used), and preserving flags meant for `docker run` (rather than the container it generates).