Forward errors from background processes spawned by pipeline's `ssh` call to the pipeline to fail if needed
The fact that we run our Spring Boot app within `nohup docker &`, which is put in the background of the parent `ssh` call, means that these multiple layers of nesting could fail to forward errors up to the grandparent CI/CD pipeline step, marking the step as a success when it in fact failed.

Adding the `ExitOnForwardFailure` SSH config option will help ensure errors are forwarded up to the parent even if `ssh` or it's spawned process on the host are put in the background.
Note: Technically, this config option is intended for port binding, which we are not doing here in our pipeline logic, but it's still a good idea to add it nonetheless.