Bug fix: Move `ssh` GH Actions command string comments outside of the string
This was an oversight -- Since the `ssh` command is being passed a string rather than the GH Actions step (which usually comments out code within `#` comments), all the content within the string was trying to be executed. This was fine a few commits ago, but ever since I improved the comments and added back-ticks in the process, this error has occurred b/c `ssh` is attempting to execute the content within the comments' back-ticks. Thus, to maintain readability but also fix this problem, move the comments within the `ssh` string outside of it to the GH Actions' code block. Also, revert temp debug code.