Bug fix: Don't force generating a new keystore there isn't one in the repo
Since it's likely that index.sh is run from within the cloned repo, that means the keystore file would never be in the directory upon app boot. This means that even if a valid one existed before, it doesn't now b/c it seems generating a new one caused issues in the previous deploy. It could be something else going on, I'll have to investigate more, but it's still superfluous regardless. Also, neither `$HOME` nor `$rootDir` were used when checking the existence of the keystore in the first place, so force using them for absolute paths.

In the process, make the keystore in `$HOME` the default one, re-order the `openssl` call to group similar flags together for minor readability improvement, and remove the useless `./index.sh certCopyToDockerContainer` call since `docker cp` only works with containers (not images).